requirements.txt
torch==2.0.0
numpy==1.24.3
requests==2.31.0
Pillow==9.4.0
matplotlib==3.8.0
scikit-learn==1.3.0
